Some logic to IcuHelpers file.
